Moved all lobby/login/shared functions into seperate files out of the cloak-server
